pyproject.toml
[project]
name = "agent-bridge"
version = "0.1.0"
description = "Bridge between chat platforms (Slack) and CLI AI agents (Claude Code)"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.21.0",
    "aiohttp>=3.9.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
agent-bridge = "agent_bridge:main_sync"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_bridge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
